backend/requirements.txt
Flask

flask-cors

python-dotenv

Geohash
python-geohash

pymysql

Flask-Mail